+++ gradle/wrapper/gradle-wrapper.properties
@@ -1,5 +1,5 @@
 #Wed Jan 08 23:48:16 GMT 2020
-distributionUrl=https\://services.gradle.org/distributions/gradle-6.3-all.zip
+distributionUrl=https\://services.gradle.org/distributions/gradle-6.4-all.zip
 distributionBase=GRADLE_USER_HOME
 distributionPath=wrapper/dists
 zipStorePath=wrapper/dists
