+++ src/test/kotlin/api/IntegrationTest.kt
@@ -1,4 +1,4 @@
-package web
+package api
 
 import domain.IdGenerator
 import io.mockk.every
